Moved functions into helper class, fstring-ed strings, remove --project parameter ...
